+++ internal/format/s3m/playback/playback.go
@@ -291,7 +291,7 @@ func (m *Manager) SetOnEffect(fn func(intf.Effect)) {
 	m.OnEffect = fn
 }
 
-func (m *Manager) SetEnvelopePosition(v uint8) {
+func (m *Manager) SetEnvelopePosition(v int) {
 }
 
 // SetupSampler configures the internal sampler
